Remove feature flag and old UI
