+++ packages/studio-base/src/AppSetting.ts
@@ -20,7 +20,6 @@ export enum AppSetting {
 
   // Experimental features
   SHOW_DEBUG_PANELS = "showDebugPanels",
-  ENABLE_NEW_TOPNAV = "enableNewTopNav",
 
   // Miscellaneous
   HIDE_SIGN_IN_PROMPT = "hideSignInPrompt",
