+++ packages/studio-base/src/components/AccountSyncGraphic.tsx
