+++ packages/studio-base/src/components/EventView.tsx
