+++ packages/studio-base/src/components/MemoryUseIndicator.tsx
@@ -6,8 +6,6 @@ import { Tooltip, Typography } from "@mui/material";
 import { makeStyles } from "tss-react/mui";
 
 import { useMemoryInfo } from "@foxglove/hooks";
-import { AppSetting } from "@foxglove/studio-base/AppSetting";
-import { useAppConfigurationValue } from "@foxglove/studio-base/hooks";
 
 const useStyles = makeStyles()((theme) => ({
   root: {
@@ -36,7 +34,6 @@ function toMB(bytes: number): number {
 function MemoryUseIndicator(): JSX.Element {
   const memoryInfo = useMemoryInfo({ refreshIntervalMs: 5000 });
   const { classes, cx } = useStyles();
-  const [enableNewTopNav = true] = useAppConfigurationValue<boolean>(AppSetting.ENABLE_NEW_TOPNAV);
 
   // If we can't load memory info (maybe not supported) then we don't show any indicator
   if (!memoryInfo) {
@@ -48,11 +45,7 @@ function MemoryUseIndicator(): JSX.Element {
   const limitMb = toMB(memoryInfo.jsHeapSizeLimit).toLocaleString();
 
   return (
-    <Tooltip
-      arrow={false}
-      title={`Used (MB): ${usedMb} / ${limitMb}`}
-      placement={enableNewTopNav ? "bottom" : "right"}
-    >
+    <Tooltip arrow={false} title={`Used (MB): ${usedMb} / ${limitMb}`} placement="bottom">
       <div className={classes.root}>
         <Typography className={cx(classes.label, classes.text)} variant="caption">
           MEM
