+++ packages/studio-base/src/components/PanelLayout.tsx
@@ -23,7 +23,6 @@ import {
 } from "react-mosaic-component";
 import { makeStyles } from "tss-react/mui";
 
-import { AppSetting } from "@foxglove/studio-base/AppSetting";
 import { EmptyPanelLayout } from "@foxglove/studio-base/components/EmptyPanelLayout";
 import EmptyState from "@foxglove/studio-base/components/EmptyState";
 import { useAppContext } from "@foxglove/studio-base/context/AppContext";
@@ -38,7 +37,6 @@ import { useExtensionCatalog } from "@foxglove/studio-base/context/ExtensionCata
 import { useLayoutManager } from "@foxglove/studio-base/context/LayoutManagerContext";
 import { usePanelCatalog } from "@foxglove/studio-base/context/PanelCatalogContext";
 import { useWorkspaceActions } from "@foxglove/studio-base/context/Workspace/useWorkspaceActions";
-import { useAppConfigurationValue } from "@foxglove/studio-base/hooks/useAppConfigurationValue";
 import { defaultPlaybackConfig } from "@foxglove/studio-base/providers/CurrentLayoutProvider/reducers";
 import { MosaicDropResult, PanelConfig } from "@foxglove/studio-base/types/panels";
 import { getPanelIdForType, getPanelTypeFromId } from "@foxglove/studio-base/util/layout";
@@ -214,7 +212,6 @@ export default function PanelLayout(): JSX.Element {
   const layoutLoading = useCurrentLayoutSelector(selectedLayoutLoadingSelector);
   const mosaicLayout = useCurrentLayoutSelector(selectedLayoutMosaicSelector);
   const registeredExtensions = useExtensionCatalog((state) => state.installedExtensions);
-  const [enableNewTopNav = true] = useAppConfigurationValue<boolean>(AppSetting.ENABLE_NEW_TOPNAV);
 
   const createNewLayout = async () => {
     const layoutData: Omit<LayoutData, "name" | "id"> = {
@@ -230,19 +227,9 @@ export default function PanelLayout(): JSX.Element {
       permission: "CREATOR_WRITE",
     });
     setSelectedLayoutId(layout.id);
-
-    if (!enableNewTopNav) {
-      openLayoutBrowser();
-    }
   };
 
   const selectExistingLayout = async () => {
-    if (!enableNewTopNav) {
-      const layouts = await layoutManager.getLayouts();
-      if (layouts[0]) {
-        setSelectedLayoutId(layouts[0].id);
-      }
-    }
     openLayoutBrowser();
   };
