+++ packages/studio-base/src/components/PlaybackControls/index.tsx
@@ -29,7 +29,6 @@ import { useCallback, useMemo, useState } from "react";
 import { makeStyles } from "tss-react/mui";
 
 import { Time, compare } from "@foxglove/rostime";
-import { AppSetting } from "@foxglove/studio-base/AppSetting";
 import { CreateEventDialog } from "@foxglove/studio-base/components/CreateEventDialog";
 import { DataSourceInfoView } from "@foxglove/studio-base/components/DataSourceInfoView";
 import EventIcon from "@foxglove/studio-base/components/EventIcon";
@@ -49,7 +48,6 @@ import {
   useWorkspaceStore,
 } from "@foxglove/studio-base/context/Workspace/WorkspaceContext";
 import { useWorkspaceActions } from "@foxglove/studio-base/context/Workspace/useWorkspaceActions";
-import { useAppConfigurationValue } from "@foxglove/studio-base/hooks";
 import { Player, PlayerPresence } from "@foxglove/studio-base/players/types";
 
 import PlaybackTimeDisplay from "./PlaybackTimeDisplay";
@@ -94,7 +92,6 @@ export default function PlaybackControls(props: {
 }): JSX.Element {
   const { play, pause, seek, isPlaying, getTimeInfo, playUntil } = props;
   const presence = useMessagePipeline(selectPresence);
-  const [enableNewTopNav = true] = useAppConfigurationValue<boolean>(AppSetting.ENABLE_NEW_TOPNAV);
 
   const { classes, cx } = useStyles();
   const repeat = useWorkspaceStore(selectPlaybackRepeat);
@@ -198,24 +195,22 @@ export default function PlaybackControls(props: {
                 onClick={toggleCreateEventDialog}
               />
             )}
-            {enableNewTopNav && (
-              <Tooltip
-                classes={{ popper: classes.popper }}
-                title={
-                  <Stack paddingY={0.75}>
-                    <DataSourceInfoView disableSource />
-                  </Stack>
-                }
-              >
-                <HoverableIconButton
-                  className={cx(classes.dataSourceInfoButton, {
-                    [classes.disabled]: disableControls,
-                  })}
-                  size="small"
-                  icon={<Info24Regular />}
-                />
-              </Tooltip>
-            )}
+            <Tooltip
+              classes={{ popper: classes.popper }}
+              title={
+                <Stack paddingY={0.75}>
+                  <DataSourceInfoView disableSource />
+                </Stack>
+              }
+            >
+              <HoverableIconButton
+                className={cx(classes.dataSourceInfoButton, {
+                  [classes.disabled]: disableControls,
+                })}
+                size="small"
+                icon={<Info24Regular />}
+              />
+            </Tooltip>
             <PlaybackTimeDisplay onSeek={seek} onPause={pause} />
           </Stack>
           <Stack direction="row" alignItems="center" gap={1}>
