+++ packages/studio-base/src/components/Sidebars/Sidebar.stories.tsx
@@ -4,16 +4,13 @@
 
 import { StoryObj } from "@storybook/react";
 import { screen, userEvent } from "@storybook/testing-library";
-import { PropsWithChildren, useEffect, useState } from "react";
+import { PropsWithChildren, useState } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 
-import { AppSetting } from "@foxglove/studio-base/AppSetting";
 import Stack from "@foxglove/studio-base/components/Stack";
-import { useAppConfigurationValue } from "@foxglove/studio-base/hooks";
 
-import Sidebars, { SidebarItem } from ".";
-import { NewSidebarItem } from "./NewSidebar";
+import { SidebarItem, Sidebars } from ".";
 
 export default {
   title: "components/NewSidebar",
@@ -45,16 +42,13 @@ const Z = () => <>{longText}</>;
 type LeftKey = "a" | "b" | "c";
 type RightKey = "x" | "y" | "z";
 
-const ITEMS = new Map<string, SidebarItem>([]);
-const BOTTOM_ITEMS = new Map<string, SidebarItem>([]);
-
-const LEFT_ITEMS = new Map<LeftKey, NewSidebarItem>([
+const LEFT_ITEMS = new Map<LeftKey, SidebarItem>([
   ["a", { title: "A", component: A }],
   ["b", { title: "B", component: B }],
   ["c", { title: "C", component: C }],
 ]);
 
-const RIGHT_ITEMS = new Map<RightKey, NewSidebarItem>([
+const RIGHT_ITEMS = new Map<RightKey, SidebarItem>([
   ["x", { title: "X", component: X }],
   ["y", { title: "Y", component: Y }],
   ["z", { title: "Z", component: Z }],
@@ -71,26 +65,15 @@ function Story({
 }): JSX.Element {
   const [selectedRightKey, setSelectedRightKey] = useState<RightKey | undefined>(defaultRightKey);
   const [selectedLeftKey, setSelectedLeftKey] = useState<LeftKey | undefined>(defaultLeftKey);
-  const [_, setAppBarEnabled] = useAppConfigurationValue<boolean>(AppSetting.ENABLE_NEW_TOPNAV);
   const [leftSidebarSize, setLeftSidebarSize] = useState<number | undefined>();
   const [rightSidebarSize, setRightSidebarSize] = useState<number | undefined>();
-
-  useEffect(() => {
-    void setAppBarEnabled(true);
-  }, [setAppBarEnabled]);
 
   return (
     <DndProvider backend={HTML5Backend}>
       <div style={{ height: "100%" }}>
         <Sidebars
-          items={ITEMS}
-          bottomItems={BOTTOM_ITEMS}
           rightItems={RIGHT_ITEMS}
           leftItems={LEFT_ITEMS}
-          selectedKey={undefined}
-          onSelectKey={() => {
-            // no-op
-          }}
           selectedRightKey={selectedRightKey}
           onSelectRightKey={setSelectedRightKey}
           selectedLeftKey={selectedLeftKey}
