+++ packages/studio-base/src/components/Sidebars/index.ts
@@ -2,4 +2,5 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-export { default as DataSourceSidebar } from "./DataSourceSidebar";
+export * from "./Sidebars";
+export type { SidebarItem } from "./Sidebar";
