+++ packages/studio-base/src/components/StudioLogsSettings/StudioLogsSettings.tsx
@@ -3,7 +3,6 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import SettingsTreeEditor from "@foxglove/studio-base/components/SettingsTreeEditor";
-import { SidebarContent } from "@foxglove/studio-base/components/SidebarContent";
 
 import { useStudioLogsSettingsTree } from "./useStudioLogsSettingsTree";
 
@@ -11,12 +10,4 @@ export function StudioLogsSettings(): JSX.Element {
   const logSettings = useStudioLogsSettingsTree();
 
   return <SettingsTreeEditor settings={logSettings} />;
-}
-
-export function StudioLogsSettingsSidebar(): JSX.Element {
-  return (
-    <SidebarContent overflow="auto" title="Studio Logs Settings" disablePadding>
-      <StudioLogsSettings />
-    </SidebarContent>
-  );
 }
