+++ packages/studio-desktop/src/renderer/Root.tsx
@@ -102,7 +102,6 @@ export default function Root(props: {
   const onMaximizeWindow = useCallback(() => nativeWindow.maximize(), [nativeWindow]);
   const onUnmaximizeWindow = useCallback(() => nativeWindow.unmaximize(), [nativeWindow]);
   const onCloseWindow = useCallback(() => nativeWindow.close(), [nativeWindow]);
-  const onReloadWindow = useCallback(() => nativeWindow.reload(), [nativeWindow]);
 
   useEffect(() => {
     const onEnterFullScreen = () => setFullScreen(true);
@@ -140,7 +139,6 @@ export default function Root(props: {
         onMaximizeWindow={onMaximizeWindow}
         onUnmaximizeWindow={onUnmaximizeWindow}
         onCloseWindow={onCloseWindow}
-        onReloadWindow={onReloadWindow}
         extraProviders={props.extraProviders}
       />
     </>
